Changed to using PR title instead of commit msg, generator was not getting branch name when checking out with tag (detached HEAD)
