+++ changelog-generator/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod config;
 pub mod parse;
+pub mod git_util;
